Factor out logic to instantiate thrift transport
This is so that implementations that derive from the hive connector
can reuse and customize the thrift transport used for talking
to the metastore
